Extra config for handling panes in tmux
Can now resize panes without using the leader key; note that this
requires turning off some MacOS defaults that map some system navigation
to the Control key + arrow combinations. The Shift + arrows will work
regardless of system config.

Also added an option to the commands that I use to make new panes so
that they are opened in the same directory as the pane that had focus
when the new pane was created. This makes it much easier to open a new
pane and run a process in the same directory (tests, web server, etc.)